Add files via upload
The provided code is a Python script that demonstrates a dog vs. cat image classification using a pre-trained InceptionV3 neural network architecture. It imports important modules for data manipulation as well as deep learning frameworks such as keras and tensorflow for building and deploying the model. Then it extracts the dataset from the zip file and loads cats and dogs images it in a separate folder . It further goes on constructing a neural network model which uses InceptionV3 as base model which does not iterate over top layers and only includes layers from the base model .Data augmentation techniques are applied to enhance the model's performance. The script further visualizes the training process through loss and accuracy plots. The trained model is saved, and its capabilities are showcased by making predictions on new images, with the aid of Grad-CAM visualization plotting the heatmap, image of heatmap representing region of interest and the original image and identifies and displays among the two categories.